reduce typesense operational bandwidth: use include_fields=id on export
- wild_typesense_reconcile.py: documented that ts_export_ids() already uses
  include_fields=id (confirmed working, was already in codebase)
- wild_reindex_after_derivative_rebuild.py: changed export URL from
  /documents/export to /documents/export?include_fields=id

Impact: Full collection export without include_fields transfers ~60 MB for
70,070 docs. With include_fields=id it transfers ~1 MB — a 60x reduction.

The active reconcile cron (wild-typesense-reconcile-v2, every 1h) and
wild_reindex_after_derivative_rebuild cron (every 30 min) were both
transferring full docs to extract only IDs. Fixed to transfer only IDs.

Verified: Typesense export endpoint supports include_fields parameter.
Confirmed via curl: include_fields=id returns 14 bytes/doc vs 897 bytes/doc
for full export.

Typesense collection: 70,070 docs
Full export size: 63 MB raw, ~13 MB gzip
ID-only export size: 1 MB raw, ~0.2 MB gzip

+++ image_qa_run.py
@@ -85,7 +85,7 @@ def main():
         SELECT g.id, g.slug, g.name, g.cover_photo_id, g.parent_gallery_id,
                p.slug as cover_photo_slug, p.thumb_url as cover_thumb_url
         FROM galleries g
-        LEFT JOIN photos p ON g.cover_photo_id = p.id
+        LEFT JOIN photos p ON g.cover_photo_id::int = p.id
         WHERE g.is_active = true
         LIMIT %s
     """
